+++ pyproject.toml
@@ -28,7 +28,7 @@ classifiers = [
 dependencies = [
     "agent-framework-core==1.0.1",
     "dspy==3.1.3",
-    "daytona>=0.164.0,<1",
+    "daytona>=0.167.0,<1",
     "hydra-core>=1.3,<2",
     "markitdown[all]>=0.1.0,<1",
     "omegaconf>=2.3,<3",
