+++ src/fleet_rlm/integrations/daytona/runtime_helpers.py
@@ -1,128 +1,28 @@
-"""Shared Daytona runtime helpers extracted from the sandbox runtime module."""
+"""Shared Daytona runtime helpers for client building and volume state."""
 
 from __future__ import annotations
 
-import atexit
 import asyncio
-import inspect
-import json
 import logging
-import os
-import re
-import subprocess
-import threading
 import time as _time
-from pathlib import Path, PurePosixPath
-from typing import Any, Awaitable, Callable, Coroutine, TypeVar, cast
+from typing import Any
 
-from fleet_rlm.runtime.content.ingestion import read_document_content
 from fleet_rlm.runtime.execution.storage_paths import mounted_storage_roots
 
+from .async_compat import _await_if_needed
 from .config import ResolvedDaytonaConfig
 from .diagnostics import DaytonaDiagnosticError, VolumeNotReadyError
 from .types import ContextSource
+from .workspace import _aensure_remote_directory
 
 _logger = logging.getLogger(__name__)
 
-DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH = PurePosixPath("/home/daytona/memory")
+DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH = __import__("pathlib").PurePosixPath(
+    "/home/daytona/memory"
+)
 
-_REMOTE_REF_RESOLUTION_TIMEOUT_S = 5
-_REMOTE_DIRECTORY_MODE = "755"
-_T = TypeVar("_T")
-
-
-class _AsyncCompatRunner:
-    """Keep one background event loop alive for sync Daytona compatibility calls."""
-
-    def __init__(self) -> None:
-        self._loop: asyncio.AbstractEventLoop | None = None
-        self._thread: threading.Thread | None = None
-        self._thread_id: int | None = None
-        self._started = threading.Event()
-        self._lock = threading.Lock()
-
-    def _thread_main(self) -> None:
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
-        self._loop = loop
-        self._thread_id = threading.get_ident()
-        self._started.set()
-        try:
-            loop.run_forever()
-        finally:
-            pending = asyncio.all_tasks(loop)
-            for task in pending:
-                task.cancel()
-            if pending:
-                loop.run_until_complete(
-                    asyncio.gather(*pending, return_exceptions=True)
-                )
-            loop.close()
-
-    def _ensure_started(self) -> asyncio.AbstractEventLoop:
-        loop = self._loop
-        thread = self._thread
-        if (
-            loop is not None
-            and not loop.is_closed()
-            and thread is not None
-            and thread.is_alive()
-        ):
-            return loop
-
-        with self._lock:
-            loop = self._loop
-            thread = self._thread
-            if (
-                loop is not None
-                and not loop.is_closed()
-                and thread is not None
-                and thread.is_alive()
-            ):
-                return loop
-
-            self._loop = None
-            self._thread_id = None
-            self._started.clear()
-            self._thread = threading.Thread(
-                target=self._thread_main,
-                name="daytona-async-compat",
-                daemon=True,
-            )
-            self._thread.start()
-
-        self._started.wait()
-        loop = self._loop
-        if loop is None or loop.is_closed():
-            raise RuntimeError("Async compatibility runner failed to start.")
-        return loop
-
-    def run(self, awaitable: Awaitable[_T]) -> _T:
-        loop = self._ensure_started()
-        if self._thread_id == threading.get_ident():
-            raise RuntimeError(
-                "Async compatibility runner cannot be called from its own loop thread."
-            )
-        future = asyncio.run_coroutine_threadsafe(
-            cast(Coroutine[Any, Any, _T], awaitable),
-            loop,
-        )
-        return future.result()
-
-    def shutdown(self) -> None:
-        loop = self._loop
-        thread = self._thread
-        if loop is None or loop.is_closed() or thread is None:
-            return
-        loop.call_soon_threadsafe(loop.stop)
-        thread.join(timeout=1)
-        self._loop = None
-        self._thread = None
-        self._thread_id = None
-
-
-_ASYNC_COMPAT_RUNNER = _AsyncCompatRunner()
-atexit.register(_ASYNC_COMPAT_RUNNER.shutdown)
+_VOLUME_READY_STATES = frozenset({"ready"})
+_VOLUME_ERROR_STATES = frozenset({"error", "failed", "deleted"})
 
 
 def _daytona_import_error(exc: ImportError) -> RuntimeError:
@@ -131,24 +31,6 @@ def _daytona_import_error(exc: ImportError) -> RuntimeError:
         "and configure DAYTONA_API_KEY / DAYTONA_API_URL before using Daytona "
         "commands. See https://www.daytona.io/docs/en/python-sdk/"
     )
-
-
-def _run_async_compat(
-    async_fn: Callable[..., Awaitable[_T]],
-    /,
-    *args: Any,
-    **kwargs: Any,
-) -> _T:
-    """Run an async implementation from sync code on one persistent loop."""
-
-    return _ASYNC_COMPAT_RUNNER.run(async_fn(*args, **kwargs))
-
-
-async def _await_if_needed(value: _T | Awaitable[_T]) -> _T:
-    if inspect.isawaitable(value):
-        awaited = await cast(Any, value)
-        return cast(_T, awaited)
-    return value
 
 
 def _build_daytona_client(config: ResolvedDaytonaConfig) -> Any:
@@ -163,10 +45,6 @@ def _build_daytona_client(config: ResolvedDaytonaConfig) -> Any:
             target=config.target,
         )
     )
-
-
-_VOLUME_READY_STATES = frozenset({"ready"})
-_VOLUME_ERROR_STATES = frozenset({"error", "failed", "deleted"})
 
 
 def _canonicalize_volume_state_token(value: Any) -> str:
@@ -213,7 +91,6 @@ def _volume_state_missing(
     handles as ready enough to continue, and several sync workflow tests rely
     on that contract.
     """
-
     if raw_state or normalized_state:
         return False
     return bool(getattr(volume, "id", None))
@@ -305,129 +182,6 @@ async def _await_volume_ready(
     )
 
 
-def _looks_like_commit(ref: str) -> bool:
-    return bool(re.fullmatch(r"[0-9a-fA-F]{7,40}", ref.strip()))
-
-
-def _safe_repo_name(repo_url: str) -> str:
-    tail = repo_url.rstrip("/").rsplit("/", 1)[-1]
-    tail = tail.removesuffix(".git")
-    cleaned = re.sub(r"[^a-zA-Z0-9._-]+", "-", tail).strip("-")
-    return cleaned or "repo"
-
-
-def _safe_workspace_name(repo_url: str | None) -> str:
-    return _safe_repo_name(repo_url) if repo_url else "daytona-workspace"
-
-
-def _safe_context_slug(value: str) -> str:
-    cleaned = re.sub(r"[^a-zA-Z0-9._-]+", "-", value).strip("-")
-    return cleaned or "context"
-
-
-def _list_remote_refs(repo_url: str) -> set[str]:
-    try:
-        completed = subprocess.run(
-            ["git", "ls-remote", "--heads", "--tags", repo_url],
-            capture_output=True,
-            check=False,
-            text=True,
-            timeout=_REMOTE_REF_RESOLUTION_TIMEOUT_S,
-        )
-    except (OSError, subprocess.SubprocessError):
-        return set()
-
-    if completed.returncode != 0:
-        return set()
-
-    refs: set[str] = set()
-    for line in completed.stdout.splitlines():
-        if "\t" not in line:
-            continue
-        _sha, raw_ref = line.split("\t", 1)
-        normalized = raw_ref.strip()
-        if normalized.startswith("refs/heads/"):
-            refs.add(normalized.removeprefix("refs/heads/"))
-            continue
-        if normalized.startswith("refs/tags/"):
-            refs.add(normalized.removeprefix("refs/tags/").removesuffix("^{}"))
-    return refs
-
-
-def _resolve_clone_ref(repo_url: str, ref: str | None) -> str | None:
-    normalized = str(ref or "").strip() or None
-    if normalized is None or _looks_like_commit(normalized) or "/" not in normalized:
-        return normalized
-
-    remote_refs = _list_remote_refs(repo_url)
-    if not remote_refs or normalized in remote_refs:
-        return normalized
-
-    segments = [segment for segment in normalized.split("/") if segment]
-    for end in range(len(segments) - 1, 0, -1):
-        candidate = "/".join(segments[:end])
-        if candidate in remote_refs:
-            return candidate
-    return normalized
-
-
-async def _aresolve_clone_ref(repo_url: str, ref: str | None) -> str | None:
-    return await asyncio.to_thread(_resolve_clone_ref, repo_url, ref)
-
-
-def _resolve_local_context_path(path: str) -> Path:
-    candidate = Path(path).expanduser()
-    if not candidate.is_absolute():
-        candidate = Path.cwd() / candidate
-    resolved = candidate.resolve()
-    if not resolved.exists():
-        raise DaytonaDiagnosticError(
-            f"Context path does not exist: {path}",
-            category="context_stage_error",
-            phase="context_stage",
-        )
-    if not os.access(resolved, os.R_OK):
-        raise DaytonaDiagnosticError(
-            f"Context path is not readable: {resolved}",
-            category="context_stage_error",
-            phase="context_stage",
-        )
-    return resolved
-
-
-async def _aget_work_dir(sandbox: Any) -> str:
-    if hasattr(sandbox, "get_work_dir"):
-        return str(await _await_if_needed(sandbox.get_work_dir()))
-    return "/workspace"
-
-
-async def _abuild_workspace_path(sandbox: Any, repo_url: str | None) -> str:
-    work_dir = await _aget_work_dir(sandbox)
-    workspace_name = _safe_workspace_name(repo_url)
-    return str(PurePosixPath(work_dir) / "workspace" / workspace_name)
-
-
-async def _aensure_remote_directory(fs: Any, remote_path: PurePosixPath) -> None:
-    directory = str(remote_path)
-    if directory and directory not in {".", "/"}:
-        await _await_if_needed(fs.create_folder(directory, _REMOTE_DIRECTORY_MODE))
-
-
-async def _aensure_remote_parent(fs: Any, remote_path: PurePosixPath) -> None:
-    await _aensure_remote_directory(fs, remote_path.parent)
-
-
-async def _aensure_workspace_root(*, sandbox: Any, workspace_path: str) -> None:
-    try:
-        await _aensure_remote_directory(sandbox.fs, PurePosixPath(workspace_path))
-    except Exception as exc:
-        raise DaytonaDiagnosticError(
-            f"Daytona workspace create failure: {exc}",
-            category="sandbox_create_clone_error",
-            phase="sandbox_create",
-        ) from exc
-
-
 async def _aensure_daytona_volume_layout(
     *,
     sandbox: Any,
@@ -442,7 +196,9 @@ async def _aensure_daytona_volume_layout(
             roots.buffers_root,
             roots.meta_root,
         ):
-            await _aensure_remote_directory(sandbox.fs, PurePosixPath(path))
+            await _aensure_remote_directory(
+                sandbox.fs, __import__("pathlib").PurePosixPath(path)
+            )
     except Exception as exc:
         raise DaytonaDiagnosticError(
             f"Daytona volume layout create failure: {exc}",
@@ -451,468 +207,16 @@ async def _aensure_daytona_volume_layout(
         ) from exc
 
 
-def _build_clone_kwargs(
-    *,
-    repo_url: str,
-    ref: str | None,
-    workspace_path: str,
-) -> dict[str, str]:
-    clone_kwargs: dict[str, str] = {"url": repo_url, "path": workspace_path}
-    if ref:
-        if _looks_like_commit(ref):
-            clone_kwargs["commit_id"] = ref
-        else:
-            clone_kwargs["branch"] = ref
-    return clone_kwargs
-
-
-async def _aclone_repo(
-    *,
-    sandbox: Any,
-    repo_url: str,
-    ref: str | None,
-    workspace_path: str,
-) -> None:
-    try:
-        await _aensure_remote_directory(
-            sandbox.fs, PurePosixPath(workspace_path).parent
-        )
-        await _await_if_needed(
-            sandbox.git.clone(
-                **_build_clone_kwargs(
-                    repo_url=repo_url,
-                    ref=ref,
-                    workspace_path=workspace_path,
-                )
-            )
-        )
-    except Exception as exc:
-        raise DaytonaDiagnosticError(
-            f"Daytona repo clone failure: {exc}",
-            category="sandbox_create_clone_error",
-            phase="repo_clone",
-        ) from exc
-
-
-async def _arun_admin_code(
-    *,
-    sandbox: Any,
-    code: str,
-    phase: str,
-    error_prefix: str,
-    category: str = "sandbox_create_clone_error",
-) -> str:
-    try:
-        result = await _await_if_needed(sandbox.process.code_run(code))
-    except Exception as exc:
-        raise DaytonaDiagnosticError(
-            f"{error_prefix}: {exc}",
-            category=category,
-            phase=phase,
-        ) from exc
-
-    exit_code = getattr(result, "exit_code", 0)
-    if exit_code:
-        detail = str(
-            getattr(result, "stderr", "")
-            or getattr(result, "result", "")
-            or getattr(result, "output", "")
-            or f"process exited with status {exit_code}"
-        )
-        raise DaytonaDiagnosticError(
-            f"{error_prefix}: {detail}",
-            category=category,
-            phase=phase,
-        )
-    return str(
-        getattr(result, "stdout", "")
-        or getattr(result, "result", "")
-        or getattr(result, "output", "")
-        or ""
-    )
-
-
-async def _areconcile_repo_checkout(
-    *,
-    sandbox: Any,
-    repo_url: str | None,
-    ref: str | None,
-    workspace_path: str,
-) -> None:
-    if repo_url is None:
-        await _aensure_workspace_root(
-            sandbox=sandbox,
-            workspace_path=workspace_path,
-        )
-        return
-
-    await _arun_admin_code(
-        sandbox=sandbox,
-        phase="repo_clone",
-        error_prefix="Daytona repo reconcile failure",
-        code=_build_repo_reconcile_script(
-            repo_url=repo_url,
-            ref=ref,
-            workspace_path=workspace_path,
-        ),
-    )
-
-
-def _build_repo_reconcile_script(
-    *,
-    repo_url: str,
-    ref: str | None,
-    workspace_path: str,
-) -> str:
-    return f"""
-import json as _json
-import pathlib as _pathlib
-import shutil as _shutil
-import subprocess as _subprocess
-
-repo_url = {repo_url!r}
-ref = {ref!r}
-workspace_path = {workspace_path!r}
-workspace = _pathlib.Path(workspace_path)
-workspace.parent.mkdir(parents=True, exist_ok=True)
-
-def _run(*args: str, check: bool = True):
-    completed = _subprocess.run(
-        list(args),
-        check=False,
-        capture_output=True,
-        text=True,
-    )
-    if check and completed.returncode != 0:
-        raise RuntimeError(
-            completed.stderr.strip()
-            or completed.stdout.strip()
-            or f"command failed: {{' '.join(args)}}"
-        )
-    return completed
-
-if not workspace.exists():
-    _run("git", "clone", repo_url, workspace_path)
-else:
-    git_dir = workspace / ".git"
-    if not git_dir.exists():
-        if any(workspace.iterdir()):
-            raise RuntimeError(
-                f"workspace exists without git metadata: {{workspace_path}}"
-            )
-        _run("git", "clone", repo_url, workspace_path)
-    else:
-        remote = _run(
-            "git",
-            "-C",
-            workspace_path,
-            "remote",
-            "get-url",
-            "origin",
-            check=False,
-        )
-        remote_url = remote.stdout.strip()
-        if remote.returncode != 0 or remote_url != repo_url:
-            _shutil.rmtree(workspace)
-            _run("git", "clone", repo_url, workspace_path)
-        else:
-            _run("git", "-C", workspace_path, "fetch", "--all", "--tags", "--prune")
-
-if ref:
-    remote_ref = f"refs/remotes/origin/{{ref}}"
-    remote_probe = _run(
-        "git",
-        "-C",
-        workspace_path,
-        "rev-parse",
-        "--verify",
-        remote_ref,
-        check=False,
-    )
-    local_probe = _run(
-        "git",
-        "-C",
-        workspace_path,
-        "rev-parse",
-        "--verify",
-        ref,
-        check=False,
-    )
-    if remote_probe.returncode == 0:
-        branch_probe = _run(
-            "git",
-            "-C",
-            workspace_path,
-            "show-ref",
-            "--verify",
-            f"refs/heads/{{ref}}",
-            check=False,
-        )
-        if branch_probe.returncode == 0:
-            _run("git", "-C", workspace_path, "checkout", "--force", ref)
-        else:
-            _run(
-                "git",
-                "-C",
-                workspace_path,
-                "checkout",
-                "--force",
-                "-B",
-                ref,
-                remote_ref,
-            )
-        _run("git", "-C", workspace_path, "reset", "--hard", remote_ref)
-    elif local_probe.returncode == 0:
-        _run("git", "-C", workspace_path, "checkout", "--force", ref)
-    else:
-        raise RuntimeError(f"requested ref is not available: {{ref}}")
-
-print(
-    _json.dumps(
-        {{
-            "repo_url": repo_url,
-            "ref": ref,
-            "workspace_path": workspace_path,
-        }},
-        ensure_ascii=False,
-    )
-)
-""".strip()
-
-
-async def _aclear_staged_context_paths(
-    *,
-    sandbox: Any,
-    workspace_path: str,
-) -> None:
-    context_root = PurePosixPath(workspace_path) / ".fleet-rlm" / "context"
-    await _arun_admin_code(
-        sandbox=sandbox,
-        phase="context_stage",
-        category="context_stage_error",
-        error_prefix="Daytona context reset failure",
-        code=f"""
-import pathlib as _pathlib
-import shutil as _shutil
-
-context_root = _pathlib.Path({str(context_root)!r})
-if context_root.exists():
-    _shutil.rmtree(context_root)
-print(str(context_root))
-""".strip(),
-    )
-
-
-async def _aupload_remote_text(
-    fs: Any, remote_path: PurePosixPath, content: str
-) -> None:
-    await _aensure_remote_parent(fs, remote_path)
-    await _await_if_needed(fs.upload_file(content.encode("utf-8"), str(remote_path)))
-
-
-async def _aread_document_content(path: Path) -> tuple[str, dict[str, Any]]:
-    text, metadata = await asyncio.to_thread(read_document_content, path)
-    return text, metadata if isinstance(metadata, dict) else {}
-
-
-def _build_staged_filename(*, source_path: Path, source_type: str) -> str:
-    return (
-        source_path.name
-        if source_type == "text"
-        else f"{source_path.name}.extracted.txt"
-    )
-
-
-async def _astage_local_file(
-    *,
-    fs: Any,
-    resolved_path: Path,
-    staged_root: PurePosixPath,
-    source_id: str,
-) -> ContextSource:
-    text, metadata = await _aread_document_content(resolved_path)
-    source_type = str(metadata.get("source_type") or "text")
-    staged_relative = staged_root / _build_staged_filename(
-        source_path=resolved_path,
-        source_type=source_type,
-    )
-    await _aupload_remote_text(fs, staged_relative, text)
-    return ContextSource(
-        source_id=source_id,
-        kind="file",
-        host_path=str(resolved_path),
-        staged_path=str(staged_relative),
-        source_type=source_type,
-        extraction_method=str(metadata.get("extraction_method") or "") or None,
-        file_count=1,
-    )
-
-
-async def _astage_local_directory(
-    *,
-    fs: Any,
-    resolved_path: Path,
-    staged_root: PurePosixPath,
-    source_id: str,
-) -> ContextSource:
-    warnings: list[str] = []
-    staged_count = 0
-    skipped_count = 0
-    extraction_methods: set[str] = set()
-    source_types: set[str] = set()
-
-    for local_file in sorted(
-        path for path in resolved_path.rglob("*") if path.is_file()
-    ):
-        relative_path = local_file.relative_to(resolved_path)
-        try:
-            text, metadata = await _aread_document_content(local_file)
-        except Exception as exc:
-            skipped_count += 1
-            warnings.append(f"Skipped {relative_path.as_posix()}: {exc}")
-            continue
-
-        source_type = str(metadata.get("source_type") or "text")
-        extraction_method = str(metadata.get("extraction_method") or "") or None
-        source_types.add(source_type)
-        if extraction_method:
-            extraction_methods.add(extraction_method)
-        destination_name = _build_staged_filename(
-            source_path=local_file,
-            source_type=source_type,
-        )
-        staged_relative = staged_root / relative_path.parent / destination_name
-        await _aupload_remote_text(fs, staged_relative, text)
-        staged_count += 1
-
-    if staged_count == 0:
-        raise DaytonaDiagnosticError(
-            f"No supported readable files found in directory: {resolved_path}",
-            category="context_stage_error",
-            phase="context_stage",
-        )
-
-    extraction_method = (
-        "mixed"
-        if len(extraction_methods) > 1
-        else next(iter(extraction_methods), None) or "directory_walk"
-    )
-    source_type = (
-        "mixed" if len(source_types) > 1 else next(iter(source_types), None) or "text"
-    )
-    return ContextSource(
-        source_id=source_id,
-        kind="directory",
-        host_path=str(resolved_path),
-        staged_path=str(staged_root),
-        source_type=source_type,
-        extraction_method=extraction_method,
-        file_count=staged_count,
-        skipped_count=skipped_count,
-        warnings=warnings,
-    )
-
-
-async def _astage_context_paths(
-    *,
-    sandbox: Any,
-    workspace_path: str,
-    context_paths: list[str] | None,
-    reset_existing: bool = False,
-) -> list[ContextSource]:
-    raw_paths = [
-        str(item).strip() for item in (context_paths or []) if str(item).strip()
-    ]
-    if reset_existing:
-        await _aclear_staged_context_paths(
-            sandbox=sandbox,
-            workspace_path=workspace_path,
-        )
-    if not raw_paths:
-        return []
-
-    fs = sandbox.fs
-    context_root = PurePosixPath(workspace_path) / ".fleet-rlm" / "context"
-    await _aensure_remote_directory(fs, context_root)
-    staged_sources: list[ContextSource] = []
-
-    for index, raw_path in enumerate(raw_paths, start=1):
-        resolved = _resolve_local_context_path(raw_path)
-        source_id = f"context-{index}"
-        staged_root = (
-            context_root
-            / f"{index:02d}-{_safe_context_slug(resolved.stem or resolved.name)}"
-        )
-        try:
-            if resolved.is_dir():
-                staged_sources.append(
-                    await _astage_local_directory(
-                        fs=fs,
-                        resolved_path=resolved,
-                        staged_root=staged_root,
-                        source_id=source_id,
-                    )
-                )
-            else:
-                staged_sources.append(
-                    await _astage_local_file(
-                        fs=fs,
-                        resolved_path=resolved,
-                        staged_root=staged_root,
-                        source_id=source_id,
-                    )
-                )
-        except DaytonaDiagnosticError:
-            raise
-        except Exception as exc:
-            raise DaytonaDiagnosticError(
-                f"Failed to stage context path '{resolved}': {exc}",
-                category="context_stage_error",
-                phase="context_stage",
-            ) from exc
-
-    manifest_path = context_root / "manifest.json"
-    await _aupload_remote_text(
-        fs,
-        manifest_path,
-        json.dumps(
-            {"context_sources": [item.to_dict() for item in staged_sources]},
-            ensure_ascii=False,
-            indent=2,
-        ),
-    )
-    return staged_sources
-
-
 __all__ = [
     "DAYTONA_PERSISTENT_VOLUME_MOUNT_PATH",
-    "_aget_work_dir",
-    "_aclone_repo",
-    "_abuild_workspace_path",
-    "_aclear_staged_context_paths",
-    "_aensure_remote_directory",
-    "_aensure_remote_parent",
-    "_aensure_workspace_root",
+    "ContextSource",
     "_aensure_daytona_volume_layout",
-    "_areconcile_repo_checkout",
-    "_aread_document_content",
-    "_arun_admin_code",
-    "_aresolve_clone_ref",
-    "_astage_context_paths",
-    "_astage_local_directory",
-    "_astage_local_file",
-    "_aupload_remote_text",
     "_await_if_needed",
     "_await_volume_ready",
-    "_build_clone_kwargs",
-    "_build_staged_filename",
     "_build_daytona_client",
+    "_canonicalize_volume_state_token",
     "_daytona_import_error",
-    "_looks_like_commit",
-    "_run_async_compat",
-    "_resolve_clone_ref",
-    "_resolve_local_context_path",
-    "_safe_context_slug",
-    "_safe_repo_name",
-    "_safe_workspace_name",
+    "_raise_if_volume_error",
+    "_volume_state_details",
+    "_volume_state_missing",
 ]
